+++ Pask.Test.dotCover/tasks/Test-NUnit.ps1
@@ -1,30 +1,35 @@
 Import-Task Test-NUnit -Package Pask.Test
-Import-Script Pask.Test.dotCover, Properties.dotCoverFilters -Package Pask.Test.dotCover
 
-# Synopsis: Run NUnit tests
-Task Test-NUnit {
-    $Assemblies = Get-TestAssemblies -TestFrameworkAssemblyName "nunit.framework"
+if ($EnableCodeCoverage -and $EnableCodeCoverage -eq $true) {
 
-    if ($Assemblies) {
-        $NUnit = Join-Path (Get-PackageDir "NUnit.ConsoleRunner") "tools\nunit3-console.exe"
-        $NUnitTestsResults = Join-Path $TestsResultsFullPath "NUnit.xml"
+    Import-Script Pask.Test.dotCover, Properties.dotCoverFilters -Package Pask.Test.dotCover
+
+    # Synopsis: Run NUnit tests
+    Task Test-NUnit {
+        $Assemblies = Get-TestAssemblies -TestFrameworkAssemblyName "nunit.framework"
+
+        if ($Assemblies) {
+            $NUnit = Join-Path (Get-PackageDir "NUnit.ConsoleRunner") "tools\nunit3-console.exe"
+            $NUnitTestsResults = Join-Path $TestsResultsFullPath "NUnit.xml"
         
-        if ($NUnitTestSelection) { 
-            $Where = "--where:`"`"$NUnitTestSelection`"`""
+            if ($NUnitTestSelection) { 
+                $Where = "--where:`"`"$NUnitTestSelection`"`""
+            }
+
+            New-Directory $TestsResultsFullPath | Out-Null
+
+            $dotCover = Get-dotCoverExe
+            $dotCoverOutput = Join-Path $TestsResultsFullPath "NUnit.dotCover.Snapshot.dcvr"
+            Remove-ItemSilently $dotCoverOutput
+            $dotCoverScope = Get-dotCoverScope $Assemblies
+
+            $NUnitAssemblies = $Assemblies -join "`"`" `"`""
+            $NUnitArguments = "--work:`"`"$TestsResultsFullPath`"`" --result:`"`"$NUnitTestsResults`"`" $Where --noheader `"`"$NUnitAssemblies`"`""
+
+            Exec { & "$dotCover" cover /TargetExecutable="$NUnit" /TargetArguments="$NUnitArguments" /Output="$dotCoverOutput" /Scope="$dotCoverScope" /Filters="`"$dotCoverFilters`"" /AttributeFilters="`"$dotCoverAttributeFilters`"" /ReturnTargetExitCode }
+        } else {
+            Write-BuildMessage "NUnit tests not found" -ForegroundColor "Yellow"
         }
-
-        New-Directory $TestsResultsFullPath | Out-Null
-
-        $dotCover = Get-dotCoverExe
-        $dotCoverOutput = Join-Path $TestsResultsFullPath "NUnit.dotCover.Snapshot.dcvr"
-        Remove-ItemSilently $dotCoverOutput
-        $dotCoverScope = Get-dotCoverScope $Assemblies
-
-        $NUnitAssemblies = $Assemblies -join "`"`" `"`""
-        $NUnitArguments = "--work:`"`"$TestsResultsFullPath`"`" --result:`"`"$NUnitTestsResults`"`" $Where --noheader `"`"$NUnitAssemblies`"`""
-
-        Exec { & "$dotCover" cover /TargetExecutable="$NUnit" /TargetArguments="$NUnitArguments" /Output="$dotCoverOutput" /Scope="$dotCoverScope" /Filters="`"$dotCoverFilters`"" /AttributeFilters="`"$dotCoverAttributeFilters`"" /ReturnTargetExitCode }
-    } else {
-        Write-BuildMessage "NUnit tests not found" -ForegroundColor "Yellow"
     }
+
 }
